fix go lint

+++ pluginmanager/flusher_wrapper_v1.go
@@ -66,6 +66,6 @@ func (wrapper *FlusherWrapperV1) Flush(projectName string, logstoreName string, 
 	if err == nil {
 		wrapper.procOutRecordsTotal.Add(int64(total))
 	}
-	wrapper.procTimeMS.Add(int64(time.Since(startTime).Milliseconds()))
+	wrapper.procTimeMS.Add(time.Since(startTime).Milliseconds())
 	return err
 }
